refactor: move ImageAnalysisResultEntitySerializer
move from Utilities to Entities folder
